override the related field between the collection modle and the product modle to make it easer to access useing related adn annotate methods

+++ storefront/store/views.py
@@ -42,7 +42,7 @@ def product_detail(request, pk):
 def collection_list(request):
     if request.method == 'GET':
         # queryset = Collection.objects.prefetch_related('product_set').all()
-        queryset = Collection.objects.annotate(products_count=Count('product')).all()
+        queryset = Collection.objects.annotate(products_count=Count('products')).all()
         serializer = CollectionSerializer(queryset, many=True)
         return Response(data=serializer.data, status=status.HTTP_200_OK)
     elif request.method == 'POST':
@@ -53,7 +53,7 @@ def collection_list(request):
         
 @api_view(['GET', 'PUT', 'DELETE'])
 def collection_detail(request, pk):
-    collection = get_object_or_404(Collection, pk=pk)
+    collection = get_object_or_404(Collection.objects.annotate(products_count=Count('products')), pk=pk)
     if request.method == 'GET':
         serializer = CollectionSerializer(collection)
         return Response(data=serializer.data, status=status.HTTP_200_OK)
@@ -63,5 +63,7 @@ def collection_detail(request, pk):
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_200_OK)
     elif request.method == 'DELETE':
+        if collection.products.count() > 0:
+            return Response({'error': 'This Collection cannot be deleted because it has products'},status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
